requirements/base.txt
asgiref==3.5.2
autopep8==1.7.0
boto3==1.24.59
botocore==1.27.59
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
dj-database-url==1.0.0
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-rest-passwordreset==1.2.1
django-storages==1.13.1
django-taggit==3.0.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
factory-boy==3.2.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumberslite==8.12.54
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.27
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
